requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
pandas==1.4.2
openpyxl==3.0.9
django-pam==2.1.0
Pillow == 8.3.2
scikit-learn == 1.0
rasterio
docker
pyshp
geopandas
matplotlib
numpy
pyinotify
django-allauth
psutil
dockerspawner
requests
djangorestframework
markdown
django-filter
django-cors-headers
celery